sql: move explainPlanNode to its own file.
